Move e2e test utils to setup.rs
